* 'master' of https://github.com/himleyb85/bitR:
  fixed #199
  kinda fixed it

+++ bmapi/views.py
@@ -12,6 +12,7 @@ from bmapi.models import *
 from pprint import pprint
 import uuid
 import json
+import time
 
 
 def login_token(request, user):
@@ -42,9 +43,10 @@ class Login( View ):
 
 class Logout( View ):
     def post(self, request ):
-        the_jason = json.loads(request.body.decode("utf-8"))
-        Token.objects.get(token=the_jason['token']).delete()
         logout( request )
+        the_jason = json.loads(request.body.decode("utf-8"))
+        if the_jason.get("token",False):
+            Token.objects.get(token=the_jason['token']).delete()
         return JsonResponse({'status':'logged out'})
 
 
@@ -93,6 +95,7 @@ def get_messages( function_name, request, chans=False ):
     if chans:
         addresses += chans
     data = [ BMclient.call( function_name, address ) for address in addresses ]
+    newlist = sorted(data[0]['data'], key=lambda k: k['receivedTime']) 
     return JsonResponse( { 'messages': data, 'chans':chans } )
